move scorecard to ts
